added txt files
