+++ message.txt
@@ -1,0 +1,1 @@
+hello world i love ruby
